tests/test_main.py
## test_main.py

import unittest
import json
from unittest.mock import patch, MagicMock
from flask import Flask

from cyber_reasoning_system.main import app

class TestMain(unittest.TestCase):

    ## Test for vulnerability_detection endpoint
    @patch('cyber_reasoning_system.main.VulnerabilityDetection')
    def test_vulnerability_detection(self, mock_vd):
        mock_vd().detect_vulnerability.return_value = {"vulnerability": "test_vulnerability"}
        with app.test_client() as client:
            response = client.post('/vulnerability_detection', data=json.dumps({'code': 'test_code'}), content_type='application/json')
            self.assertEqual(response.status_code, 200)
            self.assertEqual(response.get_json(), {"vulnerability": "test_vulnerability"})

    ## Test for patch_generation endpoint
    @patch('cyber_reasoning_system.main.PatchGeneration')
    def test_patch_generation(self, mock_pg):
        mock_pg().generate_patch.return_value = {"patch": "test_patch"}
        with app.test_client() as client:
            response = client.post('/patch_generation', data=json.dumps({'vulnerability': 'test_vulnerability'}), content_type='application/json')
            self.assertEqual(response.status_code, 200)
            self.assertEqual(response.get_json(), {"patch": "test_patch"})

    ## Test for project_management endpoint
    @patch('cyber_reasoning_system.main.ProjectManagement')
    def test_project_management(self, mock_pm):
        mock_pm().create_project.return_value = {"project": "test_project"}
        with app.test_client() as client:
            response = client.post('/project_management', data=json.dumps({'project': 'test_project'}), content_type='application/json')
            self.assertEqual(response.status_code, 200)
            self.assertEqual(response.get_json(), {"project": "test_project"})

    ## Test for scoring_evaluation endpoint
    @patch('cyber_reasoning_system.main.ScoringEvaluation')
    def test_scoring_evaluation(self, mock_se):
        mock_se().evaluate.return_value = {"score": "test_score"}
        with app.test_client() as client:
            response = client.post('/scoring_evaluation', data=json.dumps({'vulnerability': 'test_vulnerability', 'patch': 'test_patch'}), content_type='application/json')
            self.assertEqual(response.status_code, 200)
            self.assertEqual(response.get_json(), {"score": "test_score"})

    ## Test for commercialization_transition endpoint
    @patch('cyber_reasoning_system.main.CommercializationTransition')
    def test_commercialization_transition(self, mock_ct):
        mock_ct().develop_plan.return_value = {"plan": "test_plan"}
        with app.test_client() as client:
            response = client.post('/commercialization_transition', data=json.dumps({}), content_type='application/json')
            self.assertEqual(response.status_code, 200)
            self.assertEqual(response.get_json(), {"plan": "test_plan"})

if __name__ == '__main__':
    unittest.main()
